PR07: C++ Op Implementations
Part of the 22-PR split of ag_new_release_updates_2 branch.
Merge layer: 2 (native core)
Files: 841

See pr-plans/00-master-plan.md for the full split plan and merge order.

+++ libnd4j/include/ops/declarable/generic/parity_ops/zero_fraction.cpp
@@ -23,7 +23,7 @@
 #include <system/op_boilerplate.h>
 #if NOT_EXCLUDED(OP_zero_fraction)
 
-#include <ops/declarable/CustomOperations.h>
+#include <ops/declarable/headers/parity_ops.h>
 
 namespace sd {
 namespace ops {
@@ -45,7 +45,7 @@ CUSTOM_OP_IMPL(zero_fraction, 1, 1, false, 0, 0) {
   return Status::OK;
 }
 DECLARE_SHAPE_FN(zero_fraction) {
-  return SHAPELIST(ConstantShapeHelper::getInstance().scalarShapeInfo(sd::DataType::DOUBLE));
+  return SHAPELIST(ConstantShapeHelper::getInstance().scalarShapeInfo(sd::DataType::FLOAT32));
 }
 
 DECLARE_TYPES(zero_fraction) {
